Move `ImportOne` class out as `PersonIndexer`
the indexing operation is a little bit complex so deserves a separate
file
